requirements.txt
aiohttp==3.7.3
appdirs==1.4.3
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
djangorestframework-simplejwt==4.6.0
django-crispy-forms==1.11.0
django-heroku==0.3.1
djangorestframework-simplejwt==4.6.0
Django==3.1.6
dj-database-url==0.5.0
future==0.18.2
favicon==0.7.0
geoip2==4.1.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keyring==18.0.1
launchpadlib==1.10.13
django-leaflet==0.27.1
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.1.0
maxminddb==2.0.3
monotonic==1.5
more-itertools==4.2.0
multidict==5.1.0
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
Pillow==7.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycairo==1.16.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.7.3
requests==2.25.1
requests-unixsocket==0.2.0
ssh-import-id==5.10
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
whitenoise==5.2.0
